+++ src/stores/teams-store.js
@@ -1,6 +1,7 @@
+import { ref } from 'vue';
 import { defineStore } from 'pinia';
 
-// The teams store is essentially an array of team objects.
+// The teams-store holds the teams creeated by the user in a list. A team is a list of AI assistants, bots
 // Each team object has an array of team member objects and an array of messages.
 // Each team member is essentially an OpenAI assistant, with a user defined configuration.
 // A message is an objects with a role and content, e.g., { role: "user", content: "Hello"}
@@ -8,21 +9,31 @@ import { defineStore } from 'pinia';
 // The "system" message is used to set the behavior of the assistant, i.e. each team member.
 //      e.g., { role: "system", content: "You are a helpful assistant." }
 
-export const useTeamsStore = defineStore('teams', {
+export const useTeamsStore = defineStore('teams', () => {
 
-    state: () => {
-        return {
-            teams: [],
-            messages: [],
-            systemMessage: "You are a helpful assistant. You respond like you were giving examples of how to format text in markdown format using GitHub flavor.",
-            userInput: '',
-            loading: false,
-        }
-    },
+    // State properties
+    const bots = ref([]);
+    const teams = ref([]);
+    const messages = ref([]);
+    const systemMessage = ref("You are a helpful assistant. You respond like you were giving examples of how to format text in markdown format using GitHub flavor.");
+    const userInput = ref('');
+    const loading = ref(false);
 
-    actions: {
-        clearMessages() {
-            this.messages = []
-        }
+    // Actions
+    function clearMessages() {
+        messages.value = [];
     }
-})
+
+    return {
+        // State properties
+        bots,
+        teams,
+        messages,
+        systemMessage,
+        userInput,
+        loading,
+
+        // Actions
+        clearMessages
+    }
+});